// frontend/src/shared/api/generated/model/getCollectionItems200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Boardgames tracker API
 * API для проекта Boardgames tracker
 * OpenAPI spec version: 1.0.0
 */
import type { GetCollectionItems200DataItemBoardGame } from './getCollectionItems200DataItemBoardGame';
import type { GetCollectionItems200DataItemStatus } from './getCollectionItems200DataItemStatus';

export type GetCollectionItems200DataItem = {
  readonly id: number;
  readonly userId: number;
  boardGame: GetCollectionItems200DataItemBoardGame;
  /** @nullable */
  datePurchased?: string | null;
  /**
   * @minimum 0
   * @nullable
   */
  sumInRubles?: number | null;
  status: GetCollectionItems200DataItemStatus;
  /** @minimum 0 */
  readonly playCount: number;
  /**
   * @maxLength 2000
   * @nullable
   */
  comment?: string | null;
  readonly createdAt: string;
  readonly updatedAt: string;
};