// frontend/src/shared/api/generated/model/getCollectionItemsStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Boardgames tracker API
 * API для проекта Boardgames tracker
 * OpenAPI spec version: 1.0.0
 */

export type GetCollectionItemsStatus = typeof GetCollectionItemsStatus[keyof typeof GetCollectionItemsStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCollectionItemsStatus = {
  IN_COLLECTION: 'IN_COLLECTION',
  WISH_LIST: 'WISH_LIST',
  SOLD: 'SOLD',
} as const;